% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_run_bins.R
\name{fuzzy_run_bins}
\alias{fuzzy_run_bins}
\title{Run fuzzy binning on crosspoints output}
\usage{
fuzzy_run_bins(input_path, output_path, min_bin_size = 5e+05)
}
\arguments{
\item{input_path}{Path to .crosp.csv (output of crosspoints)}

\item{output_path}{Path to save .bin.csv}

\item{min_bin_size}{Minimum bin size in bp (default 500,000)}
}
\description{
Run fuzzy binning on crosspoints output
}
