% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_run_crosspoints.R
\name{fuzzy_run_crosspoints}
\alias{fuzzy_run_crosspoints}
\title{Run fuzzy crosspoints on crosspoints data with manual min_length and min_bin_size}
\usage{
fuzzy_run_crosspoints(
  input_path,
  output_path,
  min_length = 5000,
  min_bin_size = 5e+05,
  seq_depth = 10
)
}
\arguments{
\item{input_path}{Path to the input .tsv file}

\item{output_path}{Path to the output .crosp.csv file}

\item{min_length}{Manually set minimum length (default = 5000)}

\item{min_bin_size}{Manually set minimum bin size (default = 500000)}

\item{seq_depth}{Average sequencing depth (e.g., 10 for 10X). Affects HMM confidence. Default 10.}
}
\description{
Run fuzzy crosspoints on crosspoints data with manual min_length and min_bin_size
}
